Rename 'mocks' packages to 'testhelper'
They aren't mocks
(https://martinfowler.com/articles/mocksArentStubs.html#TheDifferenceBetweenMocksAndStubs),
though they were at one point in time (with Update in place). Don't use
a specific style for the test package, just use general term.
